Apply code review fixes
