+++ client/internal/wssender_cwebsocket.go
@@ -18,7 +18,9 @@ import (
 
 const (
 	defaultSendCloseMessageTimeout = 5 * time.Second
-	defaultHeartbeatIntervalMs     = 30 * 1000
+	// defaultHeartbeatIntervalMs is the default heartbeat interval in milliseconds.
+	// 30 seconds, as required by the OpAMP specification §4.1.
+	defaultHeartbeatIntervalMs = 30 * 1000
 )
 
 // WSSender implements the WebSocket client's sending portion of OpAMP protocol.
@@ -133,8 +135,8 @@ out:
 			// If there is a pending message, we will try to send it before the connection is closed.
 			case <-s.hasPendingMessage:
 				stopCtx, cancel := context.WithTimeout(context.Background(), defaultSendCloseMessageTimeout)
-				defer cancel()
 				_ = s.sendNextMessage(stopCtx)
+				cancel()
 			default:
 			}
 			// The connection close handshake is driven by runOneCycle after the sender stops.
@@ -146,9 +148,11 @@ out:
 	close(s.stopped)
 }
 
+var emptyAgentToServer = &protobufs.AgentToServer{}
+
 func (s *WSSender) sendNextMessage(ctx context.Context) error {
 	msgToSend := s.nextMessage.PopPending()
-	if msgToSend != nil && !proto.Equal(msgToSend, &protobufs.AgentToServer{}) {
+	if msgToSend != nil && !proto.Equal(msgToSend, emptyAgentToServer) {
 		// There is a pending message and the message has some fields populated.
 		return s.sendMessage(ctx, msgToSend)
 	}
