+++ client/internal/wssender_cwebsocket_test.go
@@ -8,7 +8,6 @@ import (
 	"net"
 	"net/http"
 	"runtime"
-	"sync/atomic"
 	"syscall"
 	"testing"
 	"time"
@@ -28,32 +27,29 @@ func TestWSSenderWriteWSMessageFailure_BrokenPipe(t *testing.T) {
 	// Simulate a broken pipe by injecting a connection-reset error on the
 	// client-side transport. websocket.NetConn is not suitable here because its
 	// Close performs a graceful WebSocket close handshake (blocks up to 10 s).
-	var connShim *connResetErrConn
+	var connShim *connErr
 	transport := &http.Transport{
 		DialContext: func(ctx context.Context, network, addr string) (net.Conn, error) {
 			c, err := (&net.Dialer{}).DialContext(ctx, network, addr)
 			if err != nil {
 				return nil, err
 			}
-			connShim = &connResetErrConn{Conn: c}
+			connShim = newConnResetErrConn(c)
 			return connShim, nil
 		},
 	}
 
-	conn, _, err := websocket.Dial(
-		context.Background(),
+	conn := dialCoder(
+		context.Background(), t,
 		"ws://"+srv.Endpoint,
 		&websocket.DialOptions{HTTPClient: &http.Client{Transport: transport}},
 	)
-	require.NoError(t, err)
-	conn.SetReadLimit(-1)
-	t.Cleanup(func() { _ = conn.CloseNow() })
 
 	sender := NewSender(&sharedinternal.NopLogger{})
 	ctx, cancel := context.WithCancel(context.Background())
 	t.Cleanup(cancel)
 
-	err = sender.Start(ctx, conn)
+	err := sender.Start(ctx, conn)
 	require.NoError(t, err)
 
 	// Trigger write failure injection, then schedule a send.
@@ -67,7 +63,7 @@ func TestWSSenderWriteWSMessageFailure_BrokenPipe(t *testing.T) {
 	select {
 	case <-sender.IsStopped():
 		// Expected: sender stopped and closed s.stopped
-	case <-time.After(3 * time.Second):
+	case <-time.After(senderStopTimeout):
 		t.Fatal("sender did not close s.stopped within 3s after WriteWSMessage failure")
 	}
 
@@ -78,23 +74,16 @@ func TestWSSenderWriteWSMessageFailure_BrokenPipe(t *testing.T) {
 		"StoppingErr() should be a connection reset error, got: %v", stoppingErr)
 }
 
-type connResetErrConn struct {
-	net.Conn
-	failWrite atomic.Bool
-}
-
-func (c *connResetErrConn) Write(b []byte) (n int, err error) {
-	if c.failWrite.Load() {
-		switch runtime.GOOS {
-		case "windows":
-			// wsaECONNABORTED (10053) is already caught by isConnectionResetError
-			// on Windows; WSAECONNRESET (10054) is not in the cross-platform syscall pkg.
-			return 0, &net.OpError{Op: "write", Err: syscall.Errno(wsaECONNABORTED)}
-		default:
-			return 0, &net.OpError{Op: "write", Err: syscall.Errno(syscall.ECONNRESET)}
-		}
+func newConnResetErrConn(c net.Conn) *connErr {
+	var errno syscall.Errno
+	if runtime.GOOS == "windows" {
+		// wsaECONNABORTED (10053) is already caught by isConnectionResetError
+		// on Windows; WSAECONNRESET (10054) is not in the cross-platform syscall pkg.
+		errno = syscall.Errno(wsaECONNABORTED)
+	} else {
+		errno = syscall.ECONNRESET
 	}
-	return c.Conn.Write(b)
+	return &connErr{Conn: c, writeErr: &net.OpError{Op: "write", Err: errno}}
 }
 
 func TestWSSenderWriteWSMessageFailure_ConnectionTimeout(t *testing.T) {
@@ -114,36 +103,31 @@ func TestWSSenderWriteWSMessageFailure_ConnectionTimeout(t *testing.T) {
 			return c, nil
 		},
 	}
-	conn, _, err := websocket.Dial(
-		context.Background(),
+	conn := dialCoder(
+		context.Background(), t,
 		"ws://"+srv.Endpoint,
-		&websocket.DialOptions{
-			HTTPClient: &http.Client{Transport: transport},
-		},
+		&websocket.DialOptions{HTTPClient: &http.Client{Transport: transport}},
 	)
-	require.NoError(t, err)
-	conn.SetReadLimit(-1)
-	t.Cleanup(func() { _ = conn.CloseNow() })
 
 	sender := NewSender(&sharedinternal.NopLogger{})
 	ctx, cancel := context.WithCancel(context.Background())
 	t.Cleanup(cancel)
 
-	err = sender.Start(ctx, conn)
+	err := sender.Start(ctx, conn)
 	require.NoError(t, err)
 
 	// Set a write deadline in the past on the raw TCP connection.
 	_ = rawConn.SetWriteDeadline(time.Now().Add(-1 * time.Second))
 
 	sender.NextMessage().Update(func(msg *protobufs.AgentToServer) {
-		msg.InstanceUid = make([]byte, 1024*1024)
+		msg.InstanceUid = make([]byte, largeMessageSize)
 	})
 	sender.ScheduleSend()
 
 	select {
 	case <-sender.IsStopped():
 		t.Log("Sender stopped successfully")
-	case <-time.After(3 * time.Second):
+	case <-time.After(senderStopTimeout):
 		t.Fatal("sender did not stop within 3s")
 	}
 
@@ -162,34 +146,29 @@ func TestWSSenderWriteWSMessageFailure_ConnAborted(t *testing.T) {
 
 	// Custom transport to inject our wrapped connection to simulate this case, since this case is usually hit when
 	// some local software in the system messes with the connection
-	var connShim *connAbortedErrConn
+	var connShim *connErr
 	transport := &http.Transport{
 		DialContext: func(ctx context.Context, network, addr string) (net.Conn, error) {
 			c, err := (&net.Dialer{}).DialContext(ctx, network, addr)
 			if err != nil {
 				return nil, err
 			}
-			connShim = &connAbortedErrConn{Conn: c}
+			connShim = newConnAbortedErrConn(c)
 			return connShim, nil
 		},
 	}
 
-	conn, _, err := websocket.Dial(
-		context.Background(),
+	conn := dialCoder(
+		context.Background(), t,
 		"ws://"+srv.Endpoint,
-		&websocket.DialOptions{
-			HTTPClient: &http.Client{Transport: transport},
-		},
+		&websocket.DialOptions{HTTPClient: &http.Client{Transport: transport}},
 	)
-	require.NoError(t, err)
-	conn.SetReadLimit(-1)
-	t.Cleanup(func() { _ = conn.CloseNow() })
 
 	sender := NewSender(&sharedinternal.NopLogger{})
 	ctx, cancel := context.WithCancel(context.Background())
 	t.Cleanup(cancel)
 
-	err = sender.Start(ctx, conn)
+	err := sender.Start(ctx, conn)
 	require.NoError(t, err)
 
 	// Enable error injection
@@ -203,7 +182,7 @@ func TestWSSenderWriteWSMessageFailure_ConnAborted(t *testing.T) {
 	select {
 	case <-sender.IsStopped():
 		// Expected: sender stopped
-	case <-time.After(3 * time.Second):
+	case <-time.After(senderStopTimeout):
 		t.Fatal("sender did not stop within 3s")
 	}
 
@@ -212,27 +191,4 @@ func TestWSSenderWriteWSMessageFailure_ConnAborted(t *testing.T) {
 	var opErr *net.OpError
 	require.True(t, errors.As(stoppingErr, &opErr))
 	require.True(t, opErr.Err == syscall.ECONNABORTED || opErr.Err == syscall.Errno(wsaECONNABORTED))
-}
-
-type connAbortedErrConn struct {
-	net.Conn
-	failWrite atomic.Bool
-}
-
-func (c *connAbortedErrConn) Write(b []byte) (n int, err error) {
-	if c.failWrite.Load() {
-		switch runtime.GOOS {
-		case "windows":
-			return 0, &net.OpError{
-				Op:  "write",
-				Err: syscall.Errno(wsaECONNABORTED),
-			}
-		default:
-			return 0, &net.OpError{
-				Op:  "write",
-				Err: syscall.Errno(syscall.ECONNABORTED),
-			}
-		}
-	}
-	return c.Conn.Write(b)
 }
