+++ makefile
@@ -26,6 +26,11 @@ test:
 	$(GOCMD) test -race ./...
 	cd internal/examples && $(GOCMD) test -race ./...
 
+.PHONY: test-cwebsocket
+test-cwebsocket:
+	$(GOCMD) test -race -tags cwebsocket ./...
+	cd internal/examples && $(GOCMD) test -race -tags cwebsocket ./...
+
 .PHONY: test-with-cover
 test-with-cover:
 	$(GOACC) --output=coverage.out --ignore=protobufs ./...
